Use Int16 for machine identifiers

+++ backend/src/controlmat.Application/Common/Dto/NewWashDto.cs
@@ -4,7 +4,7 @@ namespace Controlmat.Application.Common.Dto;
 
 public class NewWashDto
 {
-    public int MachineId { get; set; }
+    public short MachineId { get; set; }
     public int StartUserId { get; set; }
     public string? StartObservation { get; set; }
     public List<ProtDto> ProtEntries { get; set; } = new();
